pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pyserial>=3.5",
    "PyQt5>=5.15.10",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "tqdm>=4.67.1",
    "pyqt5>=5.15.11",
]
